requirements.txt
python-telegram-bot==13.15
APScheduler==3.6.3
tornado==6.1
cachetools==4.2.2
pytz
tzlocal
six
certifi
